refactoring of text -> list of something functions
