+++ prospector/datamodel/advisory.py
@@ -4,13 +4,18 @@ import logging
 import re
 from dataclasses import dataclass
 from datetime import datetime
-from typing import List, Optional
+from typing import List, Optional, Tuple
 
 import requests
 from pydantic import BaseModel, Field
 
 import log.util
-from processing.constants import RELEVANT_EXTENSIONS
+from processing.natural_language_processing import (
+    extract_non_nl_terms,
+    extract_path_tokens,
+    extract_products,
+    extract_versions,
+)
 
 _logger = log.util.init_local_logger()
 
@@ -38,7 +43,7 @@ class AdvisoryRecord(BaseModel):
     from_nvd: bool = False
     nvd_rest_endpoint: str = NVD_REST_ENDPOINT
     paths: List[str] = Field(default_factory=list)
-    code_tokens: List[str] = Field(default_factory=list)
+    code_tokens: Tuple[str, ...] = Field(default_factory=tuple)
 
     def analyze(self, use_nvd: bool = False):
         self.from_nvd = use_nvd
@@ -51,8 +56,7 @@ class AdvisoryRecord(BaseModel):
         )
         self.affected_products = extract_products(self.description)
         self.paths = extract_path_tokens(self.description)
-        self.code_tokens = extract_camelcase_tokens(self.description)
-        # self.non_nl_terms =
+        self.code_tokens = extract_non_nl_terms(self.description)
 
     def _get_from_nvd(self, vuln_id: str, nvd_rest_endpoint: str = NVD_REST_ENDPOINT):
         """
@@ -91,70 +95,7 @@ class AdvisoryRecord(BaseModel):
             )
 
 
-def extract_versions(text) -> "list[str]":
-    """
-    Extract all versions mentioned in the advisory text
-    """
-    regex = r"[0-9]{1,}\.[0-9]{1,}[0-9a-z.]*"
-    result = re.findall(regex, text)
-
-    return result
-
-
-def extract_products(text) -> "list[str]":
-    """
-    Extract product names from advisory text
-    """
-    # TODO implement this properly
-    regex = r"([A-Z]+[a-z\b]+)"
-    result = list(set(re.findall(regex, text)))
-    return [p for p in result if len(p) > 2]
-
-
-def extract_camelcase_tokens(text) -> "list[str]":
-    """
-    Extract camelcase or snake_case elements from advisory text
-    """
-    regex = r"([A-Za-z]+[a-z\b]+[A-Z]+[a-z\b]+)"
-    result = list(set(re.findall(regex, text)))
-    # return list(result)
-    return [p for p in result if len(p) > 2]
-    # return["blaHlafaHlsafs"]
-
-
-def extract_path_tokens(text: str, strict_extensions: bool = False) -> List[str]:
-    """
-    Used to look for paths in the text (i.e. vulnerability description)
-
-    Input:
-        text (str)
-        strict_extensions (bool): this function will always extract tokens with (back) slashes,
-            but it will only match single file names if they have the correct extension, if this argument is True
-
-    Returns:
-        list: a list of paths that are found
-    """
-    tokens = re.split(r"\s+", text)  # split the text into words
-    tokens = [
-        token.strip(",.:;-+!?)]}'\"") for token in tokens
-    ]  # removing common punctuation marks
-    paths = []
-    for token in tokens:
-        contains_path_separators = ("\\" in token) or ("/" in token)
-        separated_with_period = "." in token
-        has_relevant_extension = token.split(".")[-1] in RELEVANT_EXTENSIONS
-        is_xml_tag = token.startswith("<")
-        is_property = token.endswith("=")
-
-        is_path = contains_path_separators or (
-            has_relevant_extension if strict_extensions else separated_with_period
-        )
-        probably_not_path = is_xml_tag or is_property
-        if is_path and not probably_not_path:
-            paths.append(token)
-    return paths
-
-
+# TODO remove is not necessary
 @dataclass
 class Reference:
     """
