+++ prospector/processing/test_natural_language_processing.py
@@ -1,4 +1,9 @@
-from processing.natural_language_processing import extract_non_nl_terms
+from processing.natural_language_processing import (
+    extract_cve_references,
+    extract_jira_references,
+    extract_non_nl_terms,
+    extract_path_tokens,
+)
 
 
 def test_extract_non_nl_terms():
@@ -12,7 +17,7 @@ def test_extract_non_nl_terms():
         "the O field."
     )
 
-    terms = tuple(extract_non_nl_terms(description))
+    terms = extract_non_nl_terms(description)
 
     assert terms == (
         "org.apache.http.conn.ssl.AbstractVerifier",
@@ -31,3 +36,57 @@ def test_extract_non_nl_terms():
         "(DN)",
         '"foo,CN=www.apache.org"',
     )
+
+
+def test_adv_record_path_extraction_no_real_paths():
+    result = extract_path_tokens(ADVISORY_TEXT)
+
+    assert result == ["15.26.1", '\\"Radio', "protection,\\"]
+
+
+ADVISORY_TEXT = """Unspecified vulnerability in Uconnect before 15.26.1, as used
+    in certain Fiat Chrysler Automobiles (FCA) from 2013 to 2015 models, allows
+    remote attackers in the same cellular network to control vehicle movement,
+    cause human harm or physical damage, or modify dashboard settings via
+    vectors related to modification of entertainment-system firmware and access
+    of the CAN bus due to insufficient \\"Radio security protection,\\" as
+    demonstrated on a 2014 Jeep Cherokee Limited FWD."""
+
+ADVISORY_TEXT_2 = """In Apache Commons IO before 2.7, When invoking the method FileNameUtils.normalize
+with an improper input string, like "//../foo", or "\\..\\foo", the result would be
+the same value, thus possibly providing access to files in the parent directory,
+but not further above (thus "limited" path traversal), if the calling code would
+use the result to construct a path value."""
+
+
+def test_adv_record_path_extraction_has_real_paths():
+    result = extract_path_tokens(ADVISORY_TEXT_2)
+
+    assert result == ["2.7", "FileNameUtils.normalize", "//../foo", "\\..\\foo"]
+
+
+def test_adv_record_path_extraction_strict_extensions():
+    """
+    If strict_extensions is True, it will always extract tokens with (back) slashes,
+    but it will only collect single file names if they have the correct extension.
+    """
+    result = extract_path_tokens(
+        ADVISORY_TEXT_2
+        + " Developer.gery put something here to check if foo.java and bar.cpp will be found.",
+        strict_extensions=True,
+    )
+
+    assert result == ["//../foo", "\\..\\foo", "foo.java", "bar.cpp"]
+
+
+def test_extract_cve_identifiers():
+    result = extract_cve_references(
+        "bla bla bla CVE-1234-1234567 and CVE-1234-1234, fsafasf"
+    )
+    assert result == ["CVE-1234-1234567", "CVE-1234-1234"]
+
+
+def test_extract_jira_references():
+    commit_msg = "CXF-8535 - Checkstyle fix (cherry picked from commit bbcd8f2eb059848380fbe5af638fe94e3a9a5e1d)"
+
+    assert extract_jira_references(commit_msg) == ["CXF-8535"]
